third_party/blink/web_tests/editing/execCommand/5575101-3.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<p>This tests for a bug where outdenting a list item twice would incorrectly move it above the list it was outdented from. Below you should see 'two' in order with 'one' and 'three' and it should not be in a list.</p>
<script>

// Outdenting the item with "two" should cause it to become the next sibling
// of the item with "one" and preserve its id. The list "two" was in will be
// empty as a result of the outdent and thus removed. The list item with the
// id of "insert" will thus become empty and should also be removed leaving
// the item with "three" as the next sibling of the item with "two".
selection_test(
    [
      '<div contenteditable="true">',
        '<ol>',
          '<li>one</li>',
          '<li id="insert">',
            '<ol>',
              '<li id="outdent">|two</li>',
            '</ol>',
          '</li>',
          '<li>three</li>',
        '</ol>',
      '</div>',
    ],
    'Outdent',
    [
      '<div contenteditable="true">',
        '<ol>',
          '<li>one</li>',
          '<li id="outdent">|two</li>',
          '<li>three</li>',
        '</ol>',
      '</div>',
    ],
    'Outdent nested item two');

selection_test(
    [
      '<div contenteditable="true">',
        '<ol>',
          '<li>one</li>',
          '<li id="outdent">|two</li>',
          '<li>three</li>',
        '</ol>',
      '</div>',
    ],
    'Outdent',
    [
      '<div contenteditable="true">',
        '<ol>',
          '<li>one</li>',
        '</ol>',
        '|two',
        '<ol>',
          '<li>three</li>',
        '</ol>',
      '</div>',
    ],
    'Outdent unnested item two');

</script>
